refactor: thunks
